requirements.txt
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
datasets==3.5.0
openai==1.75.0
opencv-python==4.11.0.86
pandas==2.2.3
pillow==10.1
pinecone==6.0.2
python-dotenv==1.1.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
streamlit==1.28.2
torch==2.5.0
torchvision==0.20.0
tqdm==4.67.1
Wikipedia-API==0.8.1
